requirements.txt
# Core data handling
pandas>=2.0
tqdm>=4.64

# Transformers & NLP
transformers>=4.40
torch>=2.2
sentence-transformers>=2.7.0
spacy>=3.7

# For translation, paraphrasing, fill-mask
sacremoses>=0.1.1  # tokenizer helper for pipelines
protobuf>=3.20.0    # often needed for sentence-transformers

# Typo noise injection
nlpaug>=1.1.11

# Config + dotenv
PyYAML>=6.0
python-dotenv>=1.0.0

# Core utilities
regex>=2023.10.3

# If you use GPU quantization (BitsAndBytes for 8bit loading)
bitsandbytes>=0.42.0

# New dependencies for refactored augmentation
google-generativeai>=0.3.0  # Gemini API
requests>=2.31.0  # HTTP requests

# Dev/test
#pytest>=8.2
# If you have a Python wrapper for Llama